New DataGrid refactor
