+++ frontend/src/components/DataGrid/components/ErrorState.tsx
@@ -1,0 +1,66 @@
+import React from 'react';
+
+interface ErrorStateProps {
+  error: string;
+  onRetry: () => void;
+}
+
+export const ErrorState: React.FC<ErrorStateProps> = ({ error, onRetry }) => {
+  return (
+    <div className="my-8 mx-4">
+      <div className="max-w-md mx-auto bg-white border border-red-200 rounded-xl shadow-lg overflow-hidden">
+        <div className="bg-gradient-to-r from-red-50 to-red-100 px-6 py-4 border-b border-red-200">
+          <div className="flex items-center">
+            <div className="flex-shrink-0">
+              <div className="w-10 h-10 bg-red-500 rounded-full flex items-center justify-center">
+                <svg className="h-6 w-6 text-white" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-2.5L13.732 4c-.77-.833-1.964-.833-2.732 0L3.732 16.5c-.77.833.192 2.5 1.732 2.5z" />
+                </svg>
+              </div>
+            </div>
+            <div className="ml-4">
+              <h3 className="text-lg font-semibold text-red-800">Oops! Something went wrong</h3>
+              <p className="text-sm text-red-600">We couldn't load your jobs right now</p>
+            </div>
+          </div>
+        </div>
+        
+        <div className="px-6 py-4">
+          <div className="mb-4">
+            <p className="text-sm text-gray-600 leading-relaxed">
+              <span className="font-medium">Error details:</span> {error}
+            </p>
+          </div>
+          
+          <div className="flex flex-col sm:flex-row gap-3">
+            <button
+              onClick={onRetry}
+              className="flex-1 inline-flex items-center justify-center px-4 py-2.5 bg-red-600 hover:bg-red-700 text-white text-sm font-medium rounded-lg transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-red-500 focus:ring-offset-2"
+            >
+              <svg className="w-4 h-4 mr-2" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
+              </svg>
+              Try Again
+            </button>
+            
+            <button
+              onClick={() => window.location.reload()}
+              className="flex-1 inline-flex items-center justify-center px-4 py-2.5 bg-gray-100 hover:bg-gray-200 text-gray-700 text-sm font-medium rounded-lg transition-colors duration-200 focus:outline-none focus:ring-2 focus:ring-gray-500 focus:ring-offset-2"
+            >
+              <svg className="w-4 h-4 mr-2" fill="none" viewBox="0 0 24 24" stroke="currentColor">
+                <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 4v5h.582m15.356 2A8.001 8.001 0 004.582 9m0 0H9m11 11v-5h-.581m0 0a8.003 8.003 0 01-15.357-2m15.357 2H15" />
+              </svg>
+              Refresh Page
+            </button>
+          </div>
+        </div>
+        
+        <div className="bg-gray-50 px-6 py-3 border-t border-gray-200">
+          <p className="text-xs text-gray-500 text-center">
+            If the problem persists, please contact support or try again later.
+          </p>
+        </div>
+      </div>
+    </div>
+  );
+};
