+++ frontend/src/components/DataGrid/components/Pagination.tsx
@@ -1,0 +1,99 @@
+import React from 'react';
+
+interface PaginationProps {
+  currentPage: number;
+  totalPages: number;
+  totalItems: number;
+  pageSize: number;
+  onPageChange: (page: number) => void;
+  onPageSizeChange: (e: React.ChangeEvent<HTMLSelectElement>) => void;
+}
+
+export const Pagination: React.FC<PaginationProps> = ({
+  currentPage,
+  totalPages,
+  totalItems,
+  pageSize,
+  onPageChange,
+  onPageSizeChange,
+}) => {
+  return (
+    <div className="px-4 py-3 bg-white border-t border-gray-200 sm:px-6">
+      <div className="flex flex-col sm:flex-row justify-between items-center gap-4 text-sm text-gray-700">
+        <div className="flex items-center gap-2">
+          <span>Show</span>
+          <select
+            value={pageSize}
+            onChange={onPageSizeChange}
+            className="px-2 py-1 border border-gray-300 rounded-md text-sm"
+          >
+            <option value={5}>5</option>
+            <option value={10}>10</option>
+            <option value={20}>20</option>
+            <option value={50}>50</option>
+          </select>
+          <span>per page</span>
+        </div>
+        <div>
+          Showing <span className="font-medium">{(currentPage - 1) * pageSize + 1}</span> to{' '}
+          <span className="font-medium">
+            {Math.min(currentPage * pageSize, totalItems)}
+          </span>{' '}
+          of <span className="font-medium">{totalItems}</span> results
+        </div>
+        <div className="flex space-x-2">
+          <button
+            onClick={() => onPageChange(currentPage - 1)}
+            disabled={currentPage === 1}
+            className="px-3 py-1 border rounded-md bg-white text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+          >
+            Previous
+          </button>
+          <div className="flex items-center space-x-1">
+            {Array.from({ length: Math.min(5, totalPages) }, (_, i) => {
+              // Show first, last, current, and 2 around current
+              let pageNum;
+              if (totalPages <= 5) {
+                pageNum = i + 1;
+              } else if (currentPage <= 3) {
+                pageNum = i + 1;
+              } else if (currentPage >= totalPages - 2) {
+                pageNum = totalPages - 4 + i;
+              } else {
+                pageNum = currentPage - 2 + i;
+              }
+              
+              if (i === 3 && currentPage < totalPages - 3) {
+                return <span key="ellipsis" className="px-2">...</span>;
+              }
+              if (i === 1 && currentPage > 4) {
+                return <span key="ellipsis-start" className="px-2">...</span>;
+              }
+              
+              return (
+                <button
+                  key={pageNum}
+                  onClick={() => onPageChange(pageNum)}
+                  className={`w-8 h-8 rounded-md ${
+                    currentPage === pageNum
+                      ? 'bg-blue-600 text-white'
+                      : 'bg-white text-gray-700 hover:bg-gray-50'
+                  }`}
+                >
+                  {pageNum}
+                </button>
+              );
+            })}
+          </div>
+          <button
+            onClick={() => onPageChange(currentPage + 1)}
+            disabled={currentPage === totalPages || totalPages === 0}
+            className="px-3 py-1 border rounded-md bg-white text-gray-700 hover:bg-gray-50 disabled:opacity-50 disabled:cursor-not-allowed"
+          >
+            Next
+          </button>
+        </div>
+      </div>
+    </div>
+  );
+};
